Remove coms gcc warnings.

+++ utils/src/convert_sym.c
@@ -2168,6 +2168,6 @@ strtolower(char *s)
   char *p;
 
   for(p=s; *p; p++)
-    *p = tolower(*p);
+    *p = tolower((int) *p);
 
 }
